renamed function and copied to FileUtils
